Add support for google-java-format v1.8
The constructor of JavaFormatterOptions was 'public' with the 1.0 release
of google-java-format [1]. With v1.1, a Builder was introduced and the
constructor's visibility was changed to 'private' [2]. We didn't account
for this API breakage back then and continued to use the private
constructor. With v1.8, they changed the signature of the constructor [3]
and our reflective instantiation of the class no longer works.
We use the Builder now to instantiate the JavaFormatterOptions.

Also, control over whether to KEEP or REMOVE JavadocOnlyImports was
deprecated [4] and later removed [5]. We never exposed this option to
the user, but now have to switch to method overload without this
parameter.

Since v1.8 google-java-format requires a Java 11 runtime. Tests are
skipped accordingly when executed in an unsupported JVM.

[1] https://github.com/google/google-java-format/blob/google-java-format-1.0/core/src/main/java/com/google/googlejavaformat/java/JavaFormatterOptions.java#L90
[2] https://github.com/google/google-java-format/blob/google-java-format-1.1/core/src/main/java/com/google/googlejavaformat/java/JavaFormatterOptions.java#L65
[3]
https://github.com/google/google-java-format/commit/bd176704e73771d1dd582d2cd0ae2f8859ae9ca6#diff-59821ee17345a53c480ab844f916d2c3R53
[4]
https://github.com/google/google-java-format/commit/73c522a27974018d2bbb4414f486d76717f30a23#diff-a1ff13030903b61b1d4fd7b9277d5fadR81
[5] https://github.com/google/google-java-format/commit/edfc2abd6e6580e14c05b44fd5717ec105d8523f#diff-a1ff13030903b61b1d4fd7b9277d5fadL81